+++ cmd/main/server.go
@@ -11,8 +11,8 @@ import (
 	sq "github.com/elgris/sqrl"
 	"github.com/pentops/log.go/grpc_log"
 	"github.com/pentops/log.go/log"
-	"github.com/pentops/o5-deploy-aws/app"
 	"github.com/pentops/o5-deploy-aws/awsinfra"
+	"github.com/pentops/o5-deploy-aws/cf/app"
 	"github.com/pentops/o5-deploy-aws/deployer"
 	"github.com/pentops/o5-deploy-aws/github"
 	"github.com/pentops/o5-deploy-aws/localrun"
